Add ReportBuilder class
This refactors the Report class to represent a particular set of data in
time. These objects are now built by the ReportBuilder class which
handles retrieving data from Smart Survey. This condenses the csv
builder functionality into the Report class.
