+++ spec/ask_export/report_spec.rb
@@ -1,92 +1,36 @@
 RSpec.describe AskExport::Report do
-  around do |example|
-    travel_to(Time.zone.parse("2020-05-01 12:00")) { example.run }
+  let(:responses) do
+    [
+      report_response(id: 1, region: "Scotland"),
+      report_response(id: 2, status: "partial"),
+      report_response(id: 3, status: "disqualified"),
+    ]
   end
 
-  describe "#since_time" do
-    it "defaults to 10am on the previous day in the current time zone" do
-      expect(described_class.new.since_time)
-        .to eq(Time.zone.parse("2020-04-30 10:00"))
-    end
+  describe "#to_csv" do
+    it "returns the responses formatted as a CSV" do
+      report = described_class.new(responses, nil, nil)
+      fields = %i[submission_time status user_agent region]
 
-    it "can use the SINCE_TIME environment variable to specify the relative time on " \
-       "the previous day" do
-      ClimateControl.modify(SINCE_TIME: "12:16") do
-        expect(described_class.new.since_time).to eq(Time.zone.parse("2020-04-30 12:16"))
-      end
-    end
-
-    it "can be overridden with an absolute time with the SINCE_TIME environment variable" do
-      ClimateControl.modify(SINCE_TIME: "2020-01-01 12:16") do
-        expect(described_class.new.since_time).to eq(Time.zone.parse("2020-01-01 12:16"))
-      end
-    end
-
-    it "raises an ArgumentError when given a time that can't be parsed" do
-      ClimateControl.modify(SINCE_TIME: "Not a real time") do
-        expect { described_class.new.since_time }
-          .to raise_error(ArgumentError, %("Not a real time" could not be parsed as a time))
-      end
+      expect(report.to_csv(fields)).to eq(
+        "submission_time,status,user_agent,region\n" \
+        "01/05/2020 09:00:00,completed,NCSA Mosaic/3.0 (Windows 95),Scotland\n" \
+        "01/05/2020 09:00:00,partial,NCSA Mosaic/3.0 (Windows 95),\n" \
+        "01/05/2020 09:00:00,disqualified,NCSA Mosaic/3.0 (Windows 95),\n",
+      )
     end
   end
 
-  describe "#until_time" do
-    it "defaults to 10am on the current day in the current time zone" do
-      expect(described_class.new.until_time)
-        .to eq(Time.zone.parse("2020-05-01 10:00"))
-    end
+  describe "#filename" do
+    it "returns a string with the time range, recipient and extension" do
+      report = described_class.new(
+        [],
+        Time.new(2020, 4, 30, 10, 0),
+        Time.new(2020, 5, 1, 10, 0),
+      )
 
-    it "can be overridden with the UNTIL_TIME environment variable" do
-      ClimateControl.modify(UNTIL_TIME: "2020-01-01 12:16") do
-        expect(described_class.new.until_time).to eq(Time.zone.parse("2020-01-01 12:16"))
-      end
-    end
-  end
-
-  describe "#responses" do
-    it "delegates to SurveyResponseFetcher" do
-      secret_key = "secret"
-      instance = described_class.new
-      responses = [serialised_survey_response({ id: 1 })]
-      expected_responses = [report_response({ id: 1 }, secret_key)]
-
-      expect(AskExport::SurveyResponseFetcher)
-        .to receive(:call)
-        .with(instance.since_time, instance.until_time)
-        .and_return(responses)
-
-      ClimateControl.modify(SECRET_KEY: secret_key) do
-        expect(instance.responses).to eq(expected_responses)
-      end
-    end
-  end
-
-  describe "#completed_responses" do
-    it "returns only completed survey responses" do
-      secret_key = "secret"
-
-      responses = [
-        serialised_survey_response(status: "completed", id: 1),
-        serialised_survey_response(status: "partial"),
-        serialised_survey_response(status: "disqualified"),
-      ]
-
-      expected_responses = [report_response({ id: 1 }, secret_key)]
-
-      allow(AskExport::SurveyResponseFetcher)
-        .to receive(:call)
-        .and_return(responses)
-
-      ClimateControl.modify(SECRET_KEY: secret_key) do
-        expect(described_class.new.completed_responses).to eq(expected_responses)
-      end
-    end
-  end
-
-  describe "#filename_prefix" do
-    it "returns a the time range as a prefix" do
-      expect(described_class.new.filename_prefix)
-        .to eq("2020-04-30-1000-to-2020-05-01-1000")
+      expect(report.filename("recipient", "ext"))
+        .to eq("2020-04-30-1000-to-2020-05-01-1000-recipient.ext")
     end
   end
 end
